+++ lnd_methods/signer/sign_bytes.d.ts
@@ -10,6 +10,8 @@ export type SignBytesArgs = AuthenticatedLightningArgs<{
   key_index: number;
   /** Bytes To Hash and Sign Hex Encoded String */
   preimage: string;
+  /** BIP-340 Tagged Hash Tag UTF8 String, requires `schnorr` type */
+  tag?: string;
   /** Signature Type */
   type?: 'ecdsa' | 'schnorr';
 }>;
@@ -22,11 +24,15 @@ export type SignBytesResult = {
 /**
  * Sign a sha256 hash of arbitrary bytes
  *
+ * Supported signature types: `ecdsa`, `schnorr`
+ *
  * Requires LND built with `signrpc` build tag
  *
  * Requires `signer:generate` permission
  *
  * `schnorr` signature type is not supported on LND 0.15.0 and below
+ *
+ * `tag` is not supported on LND 0.17.5 and below and requires `schnorr` type
  */
 export const signBytes: AuthenticatedLightningMethod<
   SignBytesArgs,